Fix Issue 24144 - [REG2.105] Silent file name index overflow (#15604)
* Print an ICE message for Loc.filename overflows
* Increase the Loc size again, raising the file name limit from 64K
  to 4G but keeping the index indirection from #15199. The `Loc`
  size for DMD-as-a-library is 16 bytes then, and 12 bytes otherwise.

+++ compiler/src/dmd/globals.h
@@ -357,8 +357,8 @@ struct Loc
 {
 private:
     unsigned _linnum;
-    unsigned short _charnum;
-    unsigned short fileIndex;
+    unsigned _charnum;
+    unsigned fileIndex;
 public:
     static void set(bool showColumns, MessageStyle messageStyle);
